Create Creating a guessing game
I have created a program that generates a random number and challenges the user to guess it. The program will prompt the user to input their guess, compare it to the generated number, and provide feedback if the guess is too high or too low. It will continue until the user correctly guesses the number and then display the number of attempts it took to win the game.